+++ podonos/entity/__init__.py
@@ -1,4 +1,4 @@
-from podonos.entity.flash_eval import FlashEvalResult, FlashFileInfo
+from podonos.entity.flash_eval import FlashEvalResult
 from podonos.entity.verification import (
     FileVerificationResult,
     ProcessFilesResponse,
@@ -9,7 +9,6 @@ from podonos.entity.verification import (
 __all__ = [
     "FileVerificationResult",
     "FlashEvalResult",
-    "FlashFileInfo",
     "ProcessFilesResponse",
     "VerificationErrorDetail",
     "VerifyFilesResponse",